+++ src/ContentElements/Zusagen.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Schachbulle\ContaoMitgliederverwaltungBundle\ContentElements;
+namespace Schachbulle\ContaoFernschachBundle\ContentElements;
 
 class Zusagen extends \ContentElement
 {
@@ -9,7 +9,7 @@ class Zusagen extends \ContentElement
 	 * Template
 	 * @var string
 	 */
-	protected $strTemplate = 'ce_mitgliederverwaltung_zusagen';
+	protected $strTemplate = 'ce_fernschach_zusagen';
 
 	/**
 	 * Generate the module
@@ -17,35 +17,60 @@ class Zusagen extends \ContentElement
 	protected function compile()
 	{
 
-		// Person aus Datenbank laden, wenn ID übergeben wurde
-		if($this->mvturnierverwaltung_id)
+		// Turnier(e) aus Datenbank laden
+		if($this->fernschachverwaltung_id)
 		{
-			$objApplications = $this->Database->prepare("SELECT m.id AS mitglied_id, m.nachname AS nachname, m.vorname AS vorname, a.id AS bewerbung_id, a.applicationDate AS bewerbungsdatum, a.state AS status, a.promiseDate AS zusagedatum FROM tl_mitgliederverwaltung_applications AS a LEFT JOIN tl_mitgliederverwaltung AS m ON a.pid = m.id WHERE a.tournament=? AND a.state=? ORDER BY m.nachname ASC, m.vorname ASC")
-			                                  ->execute($this->mvturnierverwaltung_id, 1);
+			// Nur ein Turnier gewünscht
+			$objTurnier = \Database::getInstance()->prepare("SELECT * FROM tl_fernschach_turniere WHERE id = ?")
+			                                      ->execute($this->fernschachverwaltung_id);
+		}
+		else
+		{
+			// alle aktiven Einladungsturniere mit Startdatum in der Zukunft gewünscht
+			$objTurnier = \Database::getInstance()->prepare("SELECT * FROM tl_fernschach_turniere WHERE published = ? AND startDate >= ? AND typ = ? ORDER BY startDate ASC")
+			                                      ->execute(1, time(), 'e');
+		}
 
-			$daten = array();
-			// Datensätze gefunden
-			if($objApplications->numRows)
+		$daten = array();
+		// Datensätze gefunden
+		if($objTurnier->numRows)
+		{
+
+			while($objTurnier->next())
 			{
+				// Turnierzusagen einlesen
+				$zusagen = array();
+				$objSpieler = \Database::getInstance()->prepare("SELECT m.id AS mitglied_id, m.nachname AS nachname, m.vorname AS vorname, a.id AS bewerbung_id, a.applicationDate AS bewerbungsdatum, a.state AS status, a.promiseDate AS zusagedatum FROM tl_mitgliederverwaltung_applications AS a LEFT JOIN tl_mitgliederverwaltung AS m ON a.pid = m.id WHERE a.tournament=? AND a.state=? ORDER BY m.nachname ASC, m.vorname ASC")
+				                                      ->execute($objTurnier->id, 1);
 
-				while($objApplications->next())
+				if($objSpieler->numRows)
 				{
-					$daten[] = array
-					(
-						'nachname' => $objApplications->nachname,
-						'vorname'  => $objApplications->vorname,
-					);
+
+					while($objSpieler->next())
+					{
+						$zusagen[] = array
+						(
+							'nachname' => $objSpieler->nachname,
+							'vorname'  => $objSpieler->vorname,
+						);
+					}
+
 				}
 
+				$daten[] = array
+				(
+					'name'            => $objTurnier->titel,
+					'applicationText' => $objTurnier->applicationText,
+					'startDate'       => date('d.m.Y', $objTurnier->startDate),
+					'spieler'         => $zusagen,
+				);
 			}
-
-			// Daten aus tl_content in das Template schreiben
-			$this->Template->daten      = $daten;
 
 		}
 
+		// Daten aus tl_content in das Template schreiben
+		$this->Template->daten = $daten;
+
 		return;
-
 	}
-
 }