+++ src/Resources/contao/languages/de/tl_content.php
@@ -1,4 +1,4 @@
 <?php 
 
 $GLOBALS['TL_LANG']['tl_content']['fernschachverwaltung_legend'] = 'Einstellungen';
-$GLOBALS['TL_LANG']['tl_content']['fernschachverwaltung_id'] = array('Turnier auswählen', 'Bitte wählen Sie das Turnier aus, zu dem die zugesagten Teilnehmer ausgegeben werden sollen.');
+$GLOBALS['TL_LANG']['tl_content']['fernschachverwaltung_id'] = array('Turnier auswählen', 'Bitte wählen Sie das Turnier aus, zu dem die zugesagten Teilnehmer ausgegeben werden sollen. Wählen Sie nichts aus, wenn Sie alle aktiven, noch nicht gestarteten Turniere anzeigen möchten.');