+++ src/Resources/contao/languages/de/tl_nc_notification.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Deutscher Schachbund e.V.
+ *
+ * @copyright  Copyright (c) 2020, redmind GmbH
+ * @author     redmind GmbH <info@redmind.de>
+ * @license    proprietary
+ */
+
+$GLOBALS['TL_LANG']['tl_nc_notification']['type']['fernschach'] = 'Fernschach-Verwaltung';
+$GLOBALS['TL_LANG']['tl_nc_notification']['type']['meldeformular'] = array('Benachrichtigung bei Absendung des Meldeformulars');